auth-service/requirements.txt
django==5.0
djangorestframework==3.15
djangorestframework-simplejwt==5.3
psycopg2-binary==2.9.7
redis==5.0.1
django-cors-headers==4.3.1
gunicorn==21.2.0